0.68
Features:
-stop akimbo sequence if there are no more enemy at the map (similar to autoshot with auto ending of battle option);
-replaced MoraleAttackSuccessNotify from OXC to BOXCE section.

Melee mechanic changes:
-added diagonal direction attacks of terrain;
-ctrl+click - forced floor hitting under unit`s feet (make floor holes, kill unconscious unit);
-alt+click - forced terrain object/decoration hitting under feet at first, then for faced tile(s).
-small/kneeled units on high stairs/highground tiles is reachable for melee;
-big units can to hit walls faced to them parallely (originally it was possible only for south unit facing direction)

+++ src/Battlescape/MeleeAttackBState.cpp
@@ -138,13 +138,17 @@ void MeleeAttackBState::init()
 		_target = _parent->getSave()->getTile(_action.target)->getUnit();
 	}
 
-	if (!_target)
+	if (!_parent->getSave()->isCtrlPressed() && !_target)
 	{
 		throw Exception("This is a known (but tricky) bug... still fixing it, sorry. In the meantime, try save scumming option or kill all aliens in debug mode to finish the mission.");
 	}
 
-	int height = _target->getFloatHeight() + (_target->getHeight() / 2) - _parent->getSave()->getTile(_action.target)->getTerrainLevel();
-	_voxel = _action.target.toVoxel() + Position(8, 8, height);
+	//int height = _target->getFloatHeight() + (_target->getHeight() / 2) - _parent->getSave()->getTile(_action.target)->getTerrainLevel();
+	//_voxel = _action.target.toVoxel() + Position(8, 8, height);
+	
+	bool isForcedMeleeToFloor = _parent->getSave()->isCtrlPressed() && _parent->getSave()->getSide() == FACTION_PLAYER && _unit->getFaction() == FACTION_PLAYER;
+	int height = isForcedMeleeToFloor ? 1 : _target->getFloatHeight() + (_target->getHeight() * 2 / 3) - _parent->getSave()->getTile(_action.target)->getTerrainLevel();
+	_voxel = isForcedMeleeToFloor ? _unit->getPosition().toVoxel() + Position(8, 8, height) : _action.target.toVoxel() + Position(8, 8, height);
 
 	if (!_parent->getSave()->getTile(_voxel.toTile()))
 	{
